* refactor implementaion of the Restaurant interface inside the scrapers to reduce cp #7 * minor cs-fixes
